Move CopyContext to policy
