+++ src/lib.rs
@@ -78,6 +78,6 @@ pub mod util;
 pub mod vm;
 
 pub use crate::plan::{
-    AllocationSemantics, BarrierSelector, CopyContext, Mutator, MutatorContext, Plan, TraceLocal,
+    AllocationSemantics, BarrierSelector, Mutator, MutatorContext, Plan, TraceLocal,
     TransitiveClosure,
 };
