+++ src/plan/immix/gc_work.rs
@@ -1,6 +1,7 @@
 use super::global::Immix;
 use crate::policy::space::Space;
 use crate::policy::immix::ImmixCopyContext;
+use crate::policy::copy_context::NoCopy;
 use crate::scheduler::gc_work::*;
 use crate::util::{Address, ObjectReference};
 use crate::vm::VMBinding;
@@ -70,7 +71,7 @@ impl<VM: VMBinding, const KIND: TraceKind> ProcessEdgesWork for ImmixProcessEdge
         } else {
             self.immix()
                 .common
-                .trace_object::<Self, crate::plan::global::NoCopy<VM>>(self, object)
+                .trace_object::<Self, NoCopy<VM>>(self, object)
         }
     }
