+++ src/plan/mod.rs
@@ -23,7 +23,6 @@ mod global;
 pub(crate) use global::create_mutator;
 pub(crate) use global::create_plan;
 pub use global::AllocationSemantics;
-pub use global::CopyContext;
 pub(crate) use global::GcStatus;
 pub use global::Plan;
