+++ src/policy/mod.rs
@@ -13,6 +13,10 @@
 /// memory).
 pub mod space;
 
+/// Copy context defines the thread local copy allocator for copying policies.
+/// For non-copying policies, in where a CopyContext is needed, they can use `NoCopy`.
+pub mod copy_context;
+
 pub mod copyspace;
 pub mod immix;
 pub mod immortalspace;
