+++ src/vm/object_model.rs
@@ -2,7 +2,7 @@ use atomic::Ordering;
 
 use self::specs::*;
 use crate::plan::AllocationSemantics;
-use crate::plan::CopyContext;
+use crate::policy::copy_context::CopyContext;
 use crate::util::metadata::header_metadata::HeaderMetadataSpec;
 use crate::util::{Address, ObjectReference};
 use crate::vm::VMBinding;
